Minor Code Refactor
