+++ src/Application/Features/KeyValues/Commands/Delete/DeleteKeyValueCommand.cs
@@ -38,6 +38,6 @@ public class DeleteKeyValueCommandHandler : IRequestHandler<DeleteKeyValueComman
             _context.KeyValues.Remove(item);
         }
         await _context.SaveChangesAsync(cancellationToken);
-        return Result.Success();
+        return await Result.SuccessAsync();
     }
 }
