+++ src/Application/Features/KeyValues/Queries/ByName/KeyValuesQueryByName.cs
@@ -12,7 +12,7 @@ public class KeyValuesQueryByName : IRequest<IEnumerable<KeyValueDto>>, ICacheab
 
     public string CacheKey => KeyValueCacheKey.GetCacheKey(Name);
 
-    public MemoryCacheEntryOptions? Options => KeyValueCacheKey.MemoryCacheEntryOptions;
+    public MemoryCacheEntryOptions Options => KeyValueCacheKey.MemoryCacheEntryOptions;
     public KeyValuesQueryByName(string name)
     {
         Name = name;
