+++ src/Blazor.Server.UI/Pages/SystemManagement/AuditTrails.razor
@@ -1,6 +1,6 @@
 @page "/system/audittrails"
-@using CleanArchitecture.Blazor.Application.AuditTrails.Queries.PaginationQuery
 @using CleanArchitecture.Blazor.Application.Features.AuditTrails.DTOs
+@using CleanArchitecture.Blazor.Application.Features.AuditTrails.Queries.PaginationQuery
 
 
 @attribute [Authorize(Policy = Permissions.AuditTrails.View)]
