+++ src/Blazor.Server.UI/Pages/SystemManagement/Dictionaries.razor
@@ -6,6 +6,7 @@
 @using CleanArchitecture.Blazor.Application.Features.KeyValues.DTOs
 @using CleanArchitecture.Blazor.Application.Features.KeyValues.Queries.ByName
 @using CleanArchitecture.Blazor.Application.Features.KeyValues.Queries.Export
+@using CleanArchitecture.Blazor.Application.Features.KeyValues.Queries.GetAll
 
 @attribute [Authorize(Policy = Permissions.Dictionaries.View)]
 @inject IStringLocalizer<Dictionaries> L
