+++ src/Infrastructure/Hubs/Constants.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace CleanArchitecture.Blazor.Infrastructure.Constants;
+namespace CleanArchitecture.Blazor.Infrastructure.Hubs;
 
 public static class SignalR
 {
