+++ src/Infrastructure/Persistence/Configurations/DocumentConfiguration.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
-namespace CleanArchitecture.Infrastructure.Persistence.Configurations;
+namespace CleanArchitecture.Blazor.Infrastructure.Persistence.Configurations;
 
 public class DocumentConfiguration : IEntityTypeConfiguration<Document>
 {
